Split out various delta tests
