+++ geode-for-redis/src/test/java/org/apache/geode/redis/internal/data/delta/InsertByteArrayDeltaUnitTest.java
@@ -1,0 +1,71 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package org.apache.geode.redis.internal.data.delta;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+
+import junitparams.Parameters;
+import junitparams.naming.TestCaseName;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import org.apache.geode.redis.internal.data.RedisList;
+import org.apache.geode.test.junit.runners.GeodeParamsRunner;
+
+@RunWith(GeodeParamsRunner.class)
+public class InsertByteArrayDeltaUnitTest {
+  @Test
+  @Parameters(method = "getInsertByteArrayIndexes")
+  @TestCaseName("{method}: index:{0}")
+  public void testInsertByteArrayDelta(int index) throws Exception {
+    ByteArrayOutputStream baos = new ByteArrayOutputStream();
+    DataOutputStream dos = new DataOutputStream(baos);
+    RedisList redisList = makeRedisList();
+    InsertByteArray source =
+        new InsertByteArray((byte) (redisList.getVersion() + 1), "newElement".getBytes(), index);
+
+    source.serializeTo(dos);
+
+    DataInputStream dis = new DataInputStream(new ByteArrayInputStream(baos.toByteArray()));
+    redisList.fromDelta(dis);
+
+    assertThat(redisList.llen()).isEqualTo(4);
+    assertThat(redisList.lindex(index)).isEqualTo("newElement".getBytes());
+  }
+
+  @SuppressWarnings("unused")
+  private Object[] getInsertByteArrayIndexes() {
+    return new Object[] {
+        0,
+        1,
+        2,
+        3
+    };
+  }
+
+  private RedisList makeRedisList() {
+    RedisList redisList = new RedisList();
+    redisList.applyAddByteArrayTailDelta("zero".getBytes());
+    redisList.applyAddByteArrayTailDelta("one".getBytes());
+    redisList.applyAddByteArrayTailDelta("two".getBytes());
+    return redisList;
+  }
+}
